requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
Flask-CORS==5.0.0
PyJWT=2.3.0
Pillow=9.0.1